requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
beautifulsoup4==4.12.2
python-multipart==0.0.6
jinja2==3.1.2
aiohttp==3.9.1
google-generativeai==0.3.2